Move JacksonModuleProvider from server to common
