+++ core/src/main/java/com/linecorp/armeria/common/JacksonModuleProvider.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.server;
+package com.linecorp.armeria.common;
 
 import java.util.List;
