+++ it/jackson-provider/src/test/java/com/linecorp/armeria/common/CustomJacksonModuleProvider.java
@@ -14,13 +14,15 @@
  * under the License.
  */
 
-package com.linecorp.armeria.server;
+package com.linecorp.armeria.common;
 
 import java.util.List;
 
 import com.fasterxml.jackson.databind.Module;
 import com.fasterxml.jackson.module.kotlin.KotlinModule;
 import com.google.common.collect.ImmutableList;
+
+import com.linecorp.armeria.common.JacksonModuleProvider;
 
 public final class CustomJacksonModuleProvider implements JacksonModuleProvider {
     @Override
