+++ it/jackson-provider/src/test/java/com/linecorp/armeria/common/JacksonModuleProviderTest.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.server;
+package com.linecorp.armeria.common;
 
 import static org.assertj.core.api.Assertions.assertThat;
