+++ it/jackson-provider/src/test/resources/META-INF/services/com.linecorp.armeria.common.JacksonModuleProvider
@@ -1,0 +1,1 @@
+com.linecorp.armeria.common.CustomJacksonModuleProvider
