--- it/jackson-provider/src/test/resources/META-INF/services/com.linecorp.armeria.server.JacksonModuleProvider
@@ -1,1 +1,0 @@
-com.linecorp.armeria.server.CustomJacksonModuleProvider
